stat: reorganize examples to follow example_xxx_test convention
Fixes gonum/gonum#1385.
