+++ stat/example_cca_test.go
