+++ stat/example_pca_test.go
