+++ stat/example_roc_test.go
